+++ R/client.R
@@ -3,32 +3,53 @@
 #' @export
 #' @param url (character) A url. One of \code{url} or \code{handle} required.
 #' @param opts (list) curl options
-#' @param handle A handle
+#' @param proxies an object of class \code{proxy}, as returned from the
+#' \code{\link{proxy}} function. Supports one proxy for now
+#' @param handle A handle, see \code{\link{handle}}
+#'
 #' @details
 #' \strong{Methods}
 #'   \describe{
-#'     \item{\code{get(path, query, ...)}}{
+#'     \item{\code{get(path, query, disk, stream, ...)}}{
 #'       Make a GET request
 #'     }
-#'     \item{\code{post(path, query, body, ...)}}{
+#'     \item{\code{post(path, query, body, disk, stream, ...)}}{
 #'       Make a POST request
 #'     }
-#'     \item{\code{put(path, query, body, ...)}}{
+#'     \item{\code{put(path, query, body, disk, stream, ...)}}{
 #'       Make a PUT request
 #'     }
-#'     \item{\code{patch(path, query, body, ...)}}{
+#'     \item{\code{patch(path, query, body, disk, stream, ...)}}{
 #'       Make a PATCH request
 #'     }
-#'     \item{\code{delete(path, query, body, ...)}}{
+#'     \item{\code{delete(path, query, body, disk, stream, ...)}}{
 #'       Make a DELETE request
 #'     }
-#'     \item{\code{head(path, query, ...)}}{
+#'     \item{\code{head(path, disk, stream, ...)}}{
 #'       Make a HEAD request
 #'     }
 #'   }
+#'
 #' @format NULL
 #' @usage NULL
-#' @examples \dontrun{
+#' @details Possible parameters (not all are allowed in each HTTP verb):
+#' \itemize{
+#'  \item path - URL path, appended to the base URL
+#'  \item query - query terms, as a list
+#'  \item body - body as an R list
+#'  \item encode - one of form, multipart, json, or raw
+#'  \item disk - a path to write to. if NULL (default), memory used
+#'  \item stream - an R function to determine how to stream data. if
+#'  NULL (default), memory used
+#'  \item ... curl options, only those in the acceptable set from
+#'  \code{\link[curl]{curl_options}} except the following: httpget, httppost,
+#'  post, postfields, postfieldsize, and customrequest
+#' }
+#'
+#' @seealso \code{\link{post-requests}}, \code{\link{http-headers}},
+#' \code{\link{writing-options}}
+#'
+#' @examples
 #' (x <- HttpClient$new(url = "https://httpbin.org"))
 #' x$url
 #' (res_get1 <- x$get('get'))
@@ -59,47 +80,17 @@
 #' # head request
 #' (res_head <- x$head())
 #'
-#' # set options on client initialization
-#' (res <- HttpClient$new(
-#'   url = "https://httpbin.org",
-#'   opts = list(
-#'     verbose = TRUE,
-#'     useragent = "hello world"
-#'   )
-#' ))
-#' res$opts
-#' res$get('get')
-#'
-#' # set headers
-#' (res <- HttpClient$new(
-#'   url = "https://httpbin.org",
-#'   opts = list(
-#'     verbose = TRUE
-#'   ),
-#'   headers = list(
-#'     a = "stuff",
-#'     b = "things"
-#'   )
-#' ))
-#' res$headers
-#' # reassign header value
-#' res$headers$a <- "that"
-#' # define new header
-#' res$headers$c <- "what"
-#' # request
-#' res$get('get')
-#'
-#'
-#' # handles - pass in your own handle
-#' h <- handle("https://httpbin.org")
-#' (res <- HttpClient$new(handle = h))
-#' out <- res$get("get")
-#' }
+#' # query params are URL encoded for you, so DO NOT do it yourself
+#' ## if you url encode yourself, it gets double encoded, and that's bad
+#' (x <- HttpClient$new(url = "https://httpbin.org"))
+#' res <- x$get("get", query = list(a = 'hello world'), verbose = TRUE)
+
 HttpClient <- R6::R6Class(
   'HttpClient',
   public = list(
     url = NULL,
     opts = list(),
+    proxies = list(),
     headers = list(),
     handle = NULL,
 
@@ -111,6 +102,8 @@ HttpClient <- R6::R6Class(
         cat(sprintf("    %s: %s", names(self$opts)[i],
                     self$opts[[i]]), sep = "\n")
       }
+      cat("  proxies: ", sep = "\n")
+      if (length(self$proxies)) cat(paste("    -", purl(self$proxies)), sep = "\n")
       cat("  headers: ", sep = "\n")
       for (i in seq_along(self$headers)) {
         cat(sprintf("    %s: %s", names(self$headers)[i],
@@ -119,9 +112,15 @@ HttpClient <- R6::R6Class(
       invisible(self)
     },
 
-    initialize = function(url, opts, headers, handle) {
+    initialize = function(url, opts, proxies, headers, handle) {
       if (!missing(url)) self$url <- url
       if (!missing(opts)) self$opts <- opts
+      if (!missing(proxies)) {
+        if (!inherits(proxies, "proxy")) {
+          stop("proxies input must be of class proxy", call. = FALSE)
+        }
+        self$proxies <- proxies
+      }
       if (!missing(headers)) self$headers <- headers
       if (!missing(handle)) self$handle <- handle
       if (is.null(self$url) && is.null(self$handle)) {
@@ -129,7 +128,9 @@ HttpClient <- R6::R6Class(
       }
     },
 
-    get = function(path = NULL, query = list(), ...) {
+    get = function(path = NULL, query = list(), disk = NULL,
+                   stream = NULL, ...) {
+      curl_opts_check(...)
       url <- make_url(self$url, self$handle, path, query)
       rr <- list(
         url = url,
@@ -140,32 +141,43 @@ HttpClient <- R6::R6Class(
         ),
         headers = self$headers
       )
-      rr$options <- utils::modifyList(rr$options, self$opts)
+      rr$options <- utils::modifyList(rr$options,
+                                      c(self$opts, self$proxies, ...))
+      rr$disk <- disk
+      rr$stream <- stream
       private$make_request(rr)
     },
 
-    post = function(path = NULL, query = list(), body = NULL, ...) {
+    post = function(path = NULL, query = list(), body = NULL, disk = NULL,
+                    stream = NULL, encode = "multipart", ...) {
+      curl_opts_check(...)
       url <- make_url(self$url, self$handle, path, query)
-      opts <- list(post = TRUE)
-      if (is.null(body)) {
-        opts$postfields <- raw(0)
-        opts$postfieldsize <- 0
-      }
+      # opts <- list(post = TRUE)
+      # if (is.null(body)) {
+      #   opts$postfields <- raw(0)
+      #   opts$postfieldsize <- 0
+      # }
+      opts <- prep_body(body, encode)
       rr <- list(
         url = url,
         method = "post",
-        options = c(
-          opts,
+        options = as.list(c(
+          opts$opts,
           useragent = make_ua()
-        ),
-        headers = self$headers,
-        fields = body
+        )),
+        headers = c(self$headers, opts$type),
+        fields = opts$fields
       )
-      rr$options <- utils::modifyList(rr$options, self$opts)
+      rr$options <- utils::modifyList(rr$options,
+                                      c(self$opts, self$proxies, ...))
+      rr$disk <- disk
+      rr$stream <- stream
       private$make_request(rr)
     },
 
-    put = function(path = NULL, query = list(), body = NULL, ...) {
+    put = function(path = NULL, query = list(), body = NULL, disk = NULL,
+                   stream = NULL, encode = NULL, ...) {
+      curl_opts_check(...)
       url <- make_url(self$url, self$handle, path, query)
       opts <- list(customrequest = "PUT")
       if (is.null(body)) {
@@ -182,11 +194,16 @@ HttpClient <- R6::R6Class(
         headers = self$headers,
         fields = body
       )
-      rr$options <- utils::modifyList(rr$options, self$opts)
+      rr$options <- utils::modifyList(rr$options,
+                                      c(self$opts, self$proxies, ...))
+      rr$disk <- disk
+      rr$stream <- stream
       private$make_request(rr)
     },
 
-    patch = function(path = NULL, query = list(), body = NULL, ...) {
+    patch = function(path = NULL, query = list(), body = NULL, disk = NULL,
+                     stream = NULL, encode = NULL, ...) {
+      curl_opts_check(...)
       url <- make_url(self$url, self$handle, path, query)
       opts <- list(customrequest = "PATCH")
       if (is.null(body)) {
@@ -203,11 +220,16 @@ HttpClient <- R6::R6Class(
         headers = self$headers,
         fields = body
       )
-      rr$options <- utils::modifyList(rr$options, self$opts)
+      rr$options <- utils::modifyList(rr$options,
+                                      c(self$opts, self$proxies, ...))
+      rr$disk <- disk
+      rr$stream <- stream
       private$make_request(rr)
     },
 
-    delete = function(path = NULL, query = list(), body = NULL, ...) {
+    delete = function(path = NULL, query = list(), body = NULL, disk = NULL,
+                      stream = NULL, encode = NULL, ...) {
+      curl_opts_check(...)
       url <- make_url(self$url, self$handle, path, query)
       opts <- list(customrequest = "DELETE")
       if (is.null(body)) {
@@ -224,11 +246,15 @@ HttpClient <- R6::R6Class(
         headers = self$headers,
         fields = body
       )
-      rr$options <- utils::modifyList(rr$options, self$opts)
+      rr$options <- utils::modifyList(rr$options,
+                                      c(self$opts, self$proxies, ...))
+      rr$disk <- disk
+      rr$stream <- stream
       private$make_request(rr)
     },
 
-    head = function(path = NULL, ...) {
+    head = function(path = NULL, disk = NULL, stream = NULL, ...) {
+      curl_opts_check(...)
       url <- make_url(self$url, self$handle, path, NULL)
       opts <- list(customrequest = "HEAD", nobody = TRUE)
       rr <- list(
@@ -240,7 +266,10 @@ HttpClient <- R6::R6Class(
         ),
         headers = self$headers
       )
-      rr$options <- utils::modifyList(rr$options, self$opts)
+      rr$options <- utils::modifyList(rr$options,
+                                      c(self$opts, self$proxies, ...))
+      rr$disk <- disk
+      rr$stream <- stream
       private$make_request(rr)
     }
   ),
@@ -249,6 +278,11 @@ HttpClient <- R6::R6Class(
     request = NULL,
 
     make_request = function(opts) {
+      if (xor(!is.null(opts$disk), !is.null(opts$stream))) {
+        if (!is.null(opts$disk) && !is.null(opts$stream)) {
+          stop("disk and stream can not be used together", call. = FALSE)
+        }
+      }
       curl::handle_setopt(opts$url$handle, .list = opts$options)
       if (!is.null(opts$fields)) {
         curl::handle_setform(opts$url$handle, .list = opts$fields)
@@ -267,12 +301,20 @@ HttpClient <- R6::R6Class(
         resp <- curl::curl_fetch_memory(opts$url$url, opts$url$handle)
       }
 
+      # (resp <- crul_fetch(opts))
+
       HttpResponse$new(
         method = opts$method,
         url = resp$url,
         status_code = resp$status_code,
-        request_headers = c(useragent = opts$useragent, opts$headers),
-        response_headers = curl::parse_headers(rawToChar(resp$headers)),
+        request_headers = c(useragent = opts$options$useragent, opts$headers),
+        response_headers = {
+          if (grepl("^ftp://", resp$url)) {
+            list()
+          } else {
+            headers_parse(curl::parse_headers(rawToChar(resp$headers)))
+          }
+        },
         modified = resp$modified,
         times = resp$times,
         content = resp$content,