+++ R/headers.R
@@ -1,0 +1,11 @@
+head_parse <- function(z) {
+  if (grepl("HTTP\\/", z)) {
+    list(status = z)
+  } else {
+    ff <- regexec("^([^:]*):\\s*(.*)$", z)
+    xx <- regmatches(z, ff)[[1]]
+    as.list(stats::setNames(xx[[3]], tolower(xx[[2]])))
+  }
+}
+
+headers_parse <- function(x) do.call("c", lapply(x, head_parse))